feat: create SessionManagerBase class
